refactor: split utils; move duplicated logic for Pools into useAprModal()
