+++ src/components/organisms/FarmsStakeUnstakeModal/index.tsx
@@ -10,9 +10,8 @@ import useStakeFarms from '@/hooks/farms/useStakeFarms';
 import useUnstakeFarms from '@/hooks/farms/useUnstakeFarms';
 import { useMst } from '@/store';
 import { Precisions } from '@/types';
-import { getTokenUsdPrice } from '@/utils';
+import { getTokenUsdPrice, getBalanceAmountBN } from '@/utils';
 import { BIG_ZERO, DEFAULT_TOKEN_POWER } from '@/utils/constants';
-import { getBalanceAmountBN } from '@/utils/formatters';
 import { clog, clogError } from '@/utils/logger';
 
 import './FarmsStakeUnstakeModal.scss';
