+++ src/hooks/farms/useFarmsPrices.ts
@@ -3,8 +3,8 @@ import BigNumber from 'bignumber.js/bignumber';
 import { useMst } from '@/store';
 import { Farm } from '@/types';
 import { BIG_ZERO } from '@/utils/constants';
-import { getBalanceAmount } from '@/utils/formatters';
-import { clog } from '@/utils/logger';
+import { getBalanceAmount } from '@/utils';
+// import { clog } from '@/utils/logger';
 
 export const useFarmFromLpSymbol = (lpSymbol: string): Farm => {
   const { farms } = useMst();
@@ -41,7 +41,7 @@ export const useLpTokenPrice = (symbol: string) => {
     const totalLpTokens = getBalanceAmount(new BigNumber(farm.lpTotalSupply));
     lpTokenPrice = overallValueOfAllTokensInFarm.div(totalLpTokens);
 
-    clog('TEST1', farmTokenPriceInUsd, farm.tokenAmountTotal, overallValueOfAllTokensInFarm);
+    // clog('TEST1', farmTokenPriceInUsd, farm.tokenAmountTotal, overallValueOfAllTokensInFarm);
   }
 
   return lpTokenPrice;
